requirements.txt
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pygame>=2.1.0
easyocr>=1.7.0
